requirements.txt
streamlit>=1.30.0
anthropic>=0.39.0
gspread>=6.0.0
google-auth>=2.27.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
